+++ src/main/kotlin/app/RuntimeServices.kt
@@ -7,9 +7,14 @@ import top.e404.eclean.feature.trashcan.TrashcanRepository
 import top.e404.eclean.feature.trashcan.TrashcanService
 import top.e404.eclean.feature.trashcan.TrashcanTicker
 import top.e404.eclean.platform.FoliaDetector
+import top.e404.eclean.platform.execution.ExecutionGateway
+import top.e404.eclean.platform.execution.FoliaExecutionGateway
+import top.e404.eclean.platform.execution.PaperExecutionGateway
 import top.e404.eclean.platform.FoliaSchedulerFacade
 import top.e404.eclean.platform.PaperSchedulerFacade
 import top.e404.eclean.platform.SchedulerFacade
+import top.e404.eclean.platform.runtime.RuntimePlatform
+import top.e404.eclean.platform.runtime.RuntimePlatformFactory
 import top.e404.eclean.service.StatusSnapshotService
 import top.e404.eplugin.EPlugin
 
@@ -17,7 +22,13 @@ object RuntimeServices {
     lateinit var plugin: EPlugin
         private set
 
+    lateinit var platform: RuntimePlatform
+        private set
+
     lateinit var scheduler: SchedulerFacade
+        private set
+
+    lateinit var execution: ExecutionGateway
         private set
 
     lateinit var statusSnapshots: StatusSnapshotService
@@ -43,7 +54,10 @@ object RuntimeServices {
 
     fun init(plugin: EPlugin) {
         this.plugin = plugin
-        scheduler = if (FoliaDetector.isFolia()) FoliaSchedulerFacade(plugin) else PaperSchedulerFacade(plugin)
+        val isFolia = FoliaDetector.isFolia()
+        platform = RuntimePlatformFactory.create(isFolia)
+        scheduler = if (isFolia) FoliaSchedulerFacade(plugin) else PaperSchedulerFacade(plugin)
+        execution = if (isFolia) FoliaExecutionGateway(scheduler) else PaperExecutionGateway(scheduler)
         statusSnapshots = StatusSnapshotService()
         trashcanRepository = TrashcanRepository()
         trashcanService = TrashcanService(plugin, scheduler, trashcanRepository, statusSnapshots)