+++ src/main/kotlin/platform/runtime/FoliaRuntimePlatform.kt
@@ -1,0 +1,5 @@
+package top.e404.eclean.platform.runtime
+
+data object FoliaRuntimePlatform : RuntimePlatform {
+    override val id: String = "folia"
+}