+++ src/main/kotlin/platform/runtime/PaperRuntimePlatform.kt
@@ -1,0 +1,5 @@
+package top.e404.eclean.platform.runtime
+
+data object PaperRuntimePlatform : RuntimePlatform {
+    override val id: String = "paper"
+}