+++ src/main/kotlin/platform/runtime/RuntimePlatform.kt
@@ -1,0 +1,5 @@
+package top.e404.eclean.platform.runtime
+
+interface RuntimePlatform {
+    val id: String
+}