+++ src/main/kotlin/platform/runtime/RuntimePlatformFactory.kt
@@ -1,0 +1,5 @@
+package top.e404.eclean.platform.runtime
+
+object RuntimePlatformFactory {
+    fun create(folia: Boolean): RuntimePlatform = if (folia) FoliaRuntimePlatform else PaperRuntimePlatform
+}